Turn of word wrapping for Markdown
The PandocConverter now merges adjacent String and Space elements
(see method `_importInlines`). This allows for more compact and easier
to readJSON. However, due to the way that Pandoc deals with spaces/strings
means that word wrapping can not be used. We can renable this by splitting
strings by spaces in `_exportInlines`  but we need to decide
first which convention should be used for formats like
markdown: single line paragraphs or multi-line (both have pros and
cons with respect to git diffing etc)